+++ jdi-dark-tests/src/main/java/com/epam/jdi/services/websockets/WSItemClient.java
@@ -1,13 +1,23 @@
 package com.epam.jdi.services.websockets;
 
 import com.epam.jdi.dto.Item;
-import com.epam.jdi.http.WebSocketGenericClient;
+import com.epam.jdi.http.WebSocketGenericEndpoint;
 
 import javax.websocket.ClientEndpoint;
+import javax.websocket.OnMessage;
+import javax.websocket.Session;
 
 @ClientEndpoint(
         decoders = ItemDecoder.class,
         encoders = ItemEncoder.class
 )
-public class WSItemClient extends WebSocketGenericClient<Item> {
+public class WSItemClient extends WebSocketGenericEndpoint<Item> {
+
+    @OnMessage
+    public void onMessage(Item message, Session session) {
+        logger.info("Received message");
+        lastMessage = message;
+        messages.add(lastMessage);
+        latch.countDown();
+    }
 }
